La identidad de un apoyo deja de ser el puesto que ocupa en la fila
Hasta hoy, meter un punto nuevo en medio de la línea le cambiaba el
identificador a los 23 apoyos que van detrás. No falla ruidosamente: de
ese identificador cuelgan las 99 fotos y el expediente de la falla, así
que las fotos se quedan huérfanas y la pantalla dice «no identificada».
Y `npm test` seguía verde, porque nada en el repositorio fijaba esos
identificadores.

Ahora la identidad sale del NOMBRE CANÓNICO — el que tipificó el
Ingeniero—, nunca del índice ni del nombre que grabó el GPS, que tiene
errores confirmados (el apoyo E07 quedó registrado como «E02»). Las 26
semillas posicionales de julio quedan ancladas en un registro que solo
crece, así que los 26 identificadores de producción salen byte a byte
iguales: verificado reproduciendo la fórmula vieja a mano.

Un punto intercalado entra por bisección (orden 2,5 entre el 2 y el 3)
en vez de renumerar: correr los `orden` habría movido 64 de las 99 fotos
y las habría pisado sin dejar rastro.

De regalo, un fallo que llevaba vivo desde que existe la sección 10 del
informe firmable: buscaba el apoyo del expediente por un identificador
que el levantamiento no emitía, así que imprimía siempre «Estructura no
identificada» aunque el apoyo estuviera delante.

Cuatro defectos más los cazó la auditoría adversarial con 1.014 pruebas
en verde, y se corrigieron antes de este commit: una hora real de la
cuadrilla copiada dentro de un archivo declarado sintético (L-50), la
firma del Ingeniero puesta por defecto sobre lo que él no ha confirmado,
una bisección que invertía el recorrido, y un punto que va antes del
primero colocándose al final sin aviso.

Nada sembrado todavía: falta desplegar la web con el contrato 0.5.0
—antes de sembrar, o el punto intercalado se descarta en silencio— y la
llave de administrador. El pórtico del extremo de origen queda pendiente
de verificación del Ingeniero: el sembrador lo ignora y lo dice.

1.021 pruebas, 0 fallos · contrato 0.5.0 · ADR-027 · L-50

+++ contratos/src/activos.ts
@@ -153,8 +153,22 @@ export const Apoyo = Base.extend({
    * Posición en la línea. Es lo que ORDENA los vanos. Se separa del nombre a
    * propósito: el nombre puede ser irregular (en LN-627 conviven "E022",
    * "EMP TUB" y "EMPT"), y ordenar por nombre daría vanos equivocados.
+   *
+   * NO es entero desde 0.5.0, y es a propósito. Un punto que se descubre DENTRO
+   * de un vano ya levantado (un empalme a mitad de camino) entra por BISECCIÓN:
+   * recibe el orden que hay entre sus vecinos —(2+3)/2 = 2,5— en vez de correr
+   * una posición a todos los que van detrás. Renumerar obligaría a reescribir
+   * los 26 documentos ya en producción para registrar un hecho que no cambió, y
+   * movería las fotos de apoyo, que se resuelven por este mismo campo.
+   *
+   * El cambio es ADITIVO: `z.number()` admite todo lo que admitía
+   * `z.number().int()`, así que los documentos ya escritos siguen validando sin
+   * tocarlos. Pero es de una sola dirección: un documento con `orden: 2.5` NO
+   * valida contra un bundle desplegado con el contrato anterior, y ahí no da
+   * error — se descarta EN SILENCIO. Por eso se despliega la web primero y se
+   * siembra después.
    */
-  orden: z.number().int().nonnegative(),
+  orden: z.number().nonnegative(),
   /**
    * Qué es este punto. **Decide si entra o no al cálculo.** Por defecto se
    * asume `Estructura`, pero si el levantamiento trae empalmes hay que
